Use token ranges

+++ Api/src/UrlShortener.Core/Base62EncodingExtensions.cs
@@ -7,7 +7,7 @@ public static class Base62EncodingExtensions
         "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
         "abcdefghijklmnopqrstuvwxyz";
 
-    public static string EncodeToBase62(this int number)
+    public static string EncodeToBase62(this long number)
     {
         if (number == 0)
         {
@@ -18,7 +18,7 @@ public static class Base62EncodingExtensions
 
         while (number > 0)
         {
-            result.Push(Alphanumeric[number % 62]);
+            result.Push(Alphanumeric[(int)(number % 62)]);
             number /= 62;
         }
